Add AuthRoute. Redesigned route, removed unnecessary nested routes. (Main has been removed)
