+++ src/index.js
@@ -5,9 +5,11 @@ import { Provider } from "react-redux";
 
 import store from "./redux/store";
 import Login from "./containers/Login";
-import Main from "./containers/main/Main";
 import Register from "./containers/Register";
-
+import HomePage from "./containers/HomePage";
+import AuthRoute from 'Containers/AuthRoute';
+import Chat from 'Containers//Chat';
+import Page404 from 'Containers//Page404/Page404';
 
 ReactDOM.render(
   <Provider store={store}>
@@ -15,7 +17,9 @@ ReactDOM.render(
       <Switch>
         <Route path="/login" component={Login}></Route>
         <Route path="/register" component={Register}></Route>
-        <Route path="/" component={Main}></Route>
+        <AuthRoute path="/home" component={HomePage}></AuthRoute>
+        <AuthRoute path="/chat/:userid" component={Chat}></AuthRoute>
+        <Route component={Page404}></Route>
       </Switch>
     </HashRouter>
   </Provider>,
